feat: Finalized Nav and Menu behaviors
Added a hamburger menu that responds to the same variable nState as the <app-nav> changing classes. This hamburger has a click behavior that will toggle this boolean variable to open and close the nav menu
All behaviors are in css and transitions. Transitions times and delays vary whether open or close and are thus stated in their respective classes
The nav click behavior uses the same function as when pages are clicked but passes a string variable 'hamburger-action' to change classes appropriately